Fix profileme
Previous version seemed to be trying to separate piped commands to
tally each one, but it was not actually working (the first awk
command returns only the first command; no args and no subsequent
piped commands). So for now, just simplify the command to only
tally the first of piped commands. Will try to fix later to account
for pipes.